Feature(Presentation): Сделал анимацию переходов между экранами
-Сделал анимацию переходов между экранами лента и детальное описание новости
-Отредактировал разделитель в приложении, теперь он находится сверху элемента
-Отрефакторил код
-Заменил логику обработки нажатия кнопки back во фрагментах

+++ app/src/main/java/abdulmanov/eduard/news/data/repositories/NewsRepositoryImpl.kt
@@ -1,8 +1,6 @@
 package abdulmanov.eduard.news.data.repositories
 
 import abdulmanov.eduard.news.data.db.dao.NewDao
-import abdulmanov.eduard.news.data.db.dao.NewDao.Companion.ITEMS_PER_PAGE
-import abdulmanov.eduard.news.data.db.models.NewDbModel
 import abdulmanov.eduard.news.data.network.NewsProvider
 import abdulmanov.eduard.news.domain.models.New
 import abdulmanov.eduard.news.domain.repositories.NewsRepository
@@ -13,27 +11,31 @@ class NewsRepositoryImpl(
     private val newsDao: NewDao
 ) : NewsRepository {
 
-    override fun getNews(): Single<List<New>> {
+    override fun getNews(page: Int): Single<List<New>> {
         return Single.create {
-            val freshNews = newsProvider.getNews()
-            newsDao.insertNews(freshNews)
-            newsDao.deleteOldNews()
-
-
-            val allNews = newsDao.getNewsByPage(1).map { newDbModel->
-                New(
-                    id = newDbModel.id,
-                    title = newDbModel.title,
-                    link = newDbModel.link,
-                    description = newDbModel.description,
-                    date = newDbModel.date,
-                    category = newDbModel.category,
-                    image = newDbModel.image,
-                    fullDescription = newDbModel.fullDescription
-                )
+            if (page == 1) {
+                val freshNews = newsProvider.getNews()
+                newsDao.insertNews(freshNews)
+                newsDao.deleteOldNews()
+                it.onSuccess(getNewsFromDatabase(page))
+            } else {
+                it.onSuccess(getNewsFromDatabase(page))
             }
+        }
+    }
 
-            it.onSuccess(allNews)
+    private fun getNewsFromDatabase(page: Int): List<New> {
+        return newsDao.getNewsByPage(page).map { newDbModel ->
+            New(
+                id = newDbModel.id,
+                title = newDbModel.title,
+                link = newDbModel.link,
+                description = newDbModel.description,
+                date = newDbModel.date,
+                category = newDbModel.category,
+                image = newDbModel.image,
+                fullDescription = newDbModel.fullDescription
+            )
         }
     }
 }
